+++ index.js
@@ -1,6 +1,8 @@
 /**
  * @ignore
  */
-exports.KafkaProducer = require('./lib/KafkaProducer');
-exports.KafkaConsumer = require('./lib/KafkaConsumer');
-exports.globalEvent = require('./lib/producer_pool').globalEvent;
+exports.SHKafkaProducer = require('./lib/SHKafkaProducer');
+exports.SHKafkaConsumer = require('./lib/SHKafkaConsumer');
+// exports.globalEvent = require('./lib/producer_pool').globalEvent;
+exports.RdKafkaProducer = require('./lib/RdKafkaProducer');
+exports.RdKafkaConsumer = require('./lib/RdKafkaConsumer');